Now all favorites logic @ FavoritesService
Former-commit-id: 988b2529337bd047ab52c5e03c5291a2c82946c9

+++ src/main/java/org/utn/marvellator/service/CharacterService.java
@@ -11,5 +11,5 @@ public interface CharacterService {
 
     List<MarvelCharacter> getCharacters();
 
-    MarvelCharacter getCharacterById(Long characterId);
+    MarvelCharacter getCharacterById(Integer characterId);
 }
